SELinux: A few more rules needed by libvirt for using swtpm
This patch adds yet a few more rules that are needed for running swtpm
with libvirtd. The patch includes rules for starting swtpm when QEMU
is started with svirt_tcg_t, for example when the ppc64 emulator is
started on an x86_64 host.

The additional rules were developed on Fedora 28.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>